Frontpage makover
Quite a lot to mention here...I have setup the Main Menu buttons to be
disabled if the person using the system does not have permissions. Also
the first page that gets loaded is the Main Menu. Users can log in if
they need more features than inventory search. I have placed buttons for
each of the permission levels we talked about. I have a lot of changes
to the Stores form, but I am testing stuff out on another copy of the
application. When I am happy with what it is doing with the validation,
I will post the changes.